[DNM] RMX1971: Default to 19691 for touch
When using RUI2 based recovery on RUI1 firmware, oplus_project fails to get the correct project name for RMX1971. This means it fails to load the correct firmware and touch doesn't work. To get it working on both RUI1 and RUI2, default to 19691. As this won't be used on any other devices, no issues should occur. Also default DEV_NAME to DSJM in order to load correct firmware.